Revisted layer refreshing
We had too much jitter with the old implementation which caused the LEDs
to flicker. Changed implementation so that the GSCLK determines how fast
our layers are refreshed. After the GSCLK period has finished we latch
in new data. Right after we start a new GSCLK period we can write new
data to the TLC's input buffer parallel to it completing the GSCLK
period. This scheme is way more deterministic and causes only minimal
jitter resulting in a flicker free cube.

Due to these changes we ditched all of the open LED detection. Will
introduce it at some later point in time, maybe this weekend :-)!

Reverted some kernel code, because I broke it...