RS485 is working, but...
Because the tlc5940 is quite a heavy statemachine (copying a lot of data
etc very frequently) we're experiencing some rs485 direction pin jitter,
in the order of 10s of microseconds. Unfortunately we can't use PPS to
use hardware flow control, so we might need to come up with another
solution to bring this jitter down. Not super important for now though.

Also added a target with -Wpedantic